library stuff

+++ src/routes/Library.js
@@ -2,6 +2,8 @@ import React, {Component} from 'react';
 import * as bs from 'react-bootstrap';
 
 import FAIcon from '../components/FAIcon.js';
+import DeleteModal from '../components/DeleteModal.js';
+import ImportModal from '../components/ImportModal.js';
 
 import * as songDB from '../songDB.js';
 
@@ -14,7 +16,9 @@ export default class Library extends Component {
   }
   updateSongState() {
     songDB.getAllSongs().then(songs => {
-      this.setState({...this.state, songs})
+      let allUIDs = songs.map(song => song.uid);
+      let newChecked = new Set([...this.state.checkedChildren].filter(uid => allUIDs.includes(uid)));
+      this.setState({...this.state, checkedChildren: newChecked, songs})
     });
   }
   childCheckedStateChanged(childUID, checked) {
@@ -26,15 +30,29 @@ export default class Library extends Component {
     }
     this.setState({...this.state, checkedChildren: newChecked});
   }
+  toggleAllChecks() {
+    if(this.state.checkedChildren.size) {
+      this.setState({...this.state, checkedChildren: new Set()});
+    } else {
+      let allUIDs = new Set(this.state.songs.map(song => song.uid));
+      this.setState({...this.state, checkedChildren: allUIDs});
+    }
+  }
   render() {
     return (
-      <div>
+      <>
         <h2>My library</h2>
         <bs.ListGroup>
-          {this.state.songs.map(song => (<LibraryItem key={song.uid} song={song} checkedStateChanged={this.childCheckedStateChanged.bind(this)} />))}
+          <bs.ListGroup.Item variant="secondary">
+            <bs.Form.Check
+              label={`${this.state.checkedChildren.size} selected`}
+              checked={this.state.checkedChildren.size}
+              onChange={this.toggleAllChecks.bind(this)}/>
+          </bs.ListGroup.Item>
+          {this.state.songs.map(song => (<LibraryItem key={song.uid} song={song} isChecked={this.state.checkedChildren.has(song.uid)} checkedStateChanged={this.childCheckedStateChanged.bind(this)} />))}
         </bs.ListGroup>
         <LibraryToolbar checkedSongs={this.state.checkedChildren}/>
-      </div>
+      </>
     );
   }
 }
@@ -49,16 +67,36 @@ class LibraryItem extends Component {
       <bs.ListGroup.Item action href={`/editor/${song.uid}`}>
         <bs.Form.Check
           label={song.title}
-          onClick={e => this.props.checkedStateChanged(song.uid, e.target.checked)}/>
+          checked={this.props.isChecked}
+          onChange={e => this.props.checkedStateChanged(song.uid, e.target.checked)}/>
       </bs.ListGroup.Item>
     );
   }
 }
 
 class LibraryToolbar extends Component {
-  deleteChecked() {
-    // @todo do this as 1 transaction probably
-    this.props.checkedSongs.forEach(uid => songDB.deleteSong(uid));
+  constructor() {
+    super();
+    this.state = {showDeleteModal: false, showImportModal: false};
+  }
+  showDeleteModal() {
+    this.setState({...this.state, showDeleteModal: true});
+  }
+  handleDeleteModalClose(confirmed) {
+    this.setState({...this.state, showDeleteModal: false});
+    if(confirmed) {
+      // @todo do this as 1 transaction probably
+      this.props.checkedSongs.forEach(uid => songDB.deleteSong(uid));
+    }
+  }
+  showImportModal() {
+    this.setState({...this.state, showImportModal: true});
+  }
+  handleImportModalClose(confirmed) {
+    this.setState({...this.state, showImportModal: false});
+    if(confirmed) {
+      console.log(44);
+    }
   }
   duplicateChecked() {
     // heck yeah async
@@ -70,6 +108,7 @@ class LibraryToolbar extends Component {
   }
   async downloadChecked() {
     const library = await Promise.all([...this.props.checkedSongs].map(uid => songDB.getSong(uid)));
+    library.forEach(song => delete song.uid);
     const serialized = encodeURIComponent(JSON.stringify(library));
 
     //https://ourcodeworld.com/articles/read/189/how-to-create-a-file-and-generate-a-download-with-javascript-in-the-browser-without-a-server
@@ -85,18 +124,30 @@ class LibraryToolbar extends Component {
     document.body.removeChild(element);
   }
   render() {
+    const anyChecked = !this.props.checkedSongs.size;
     return (
-      <bs.ButtonToolbar className="notochord-toolbar bg-light">
-          <bs.Button variant="danger" onClick={this.deleteChecked.bind(this)}>
-            <FAIcon icon="play-circle" /> Delete {this.props.checkedSongs.size}
-          </bs.Button>
-          <bs.Button onClick={this.duplicateChecked.bind(this)}>
-            <FAIcon icon="play-circle" /> Duplicate {this.props.checkedSongs.size}
-          </bs.Button>
-          <bs.Button onClick={this.downloadChecked.bind(this)}>
-            <FAIcon icon="play-circle" /> Download {this.props.checkedSongs.size}
-          </bs.Button>
-      </bs.ButtonToolbar>
+      <>
+        <bs.ButtonToolbar className="notochord-toolbar bg-light">
+            <bs.Button variant="danger"
+              onClick={this.showDeleteModal.bind(this)}
+              disabled={anyChecked}>
+              <FAIcon icon="trash-alt" /> Delete
+            </bs.Button>
+            <bs.Button onClick={this.duplicateChecked.bind(this)}
+              disabled={anyChecked}>
+              <FAIcon icon="copy" /> Duplicate
+            </bs.Button>
+            <bs.Button onClick={this.downloadChecked.bind(this)}
+              disabled={anyChecked}>
+              <FAIcon fastyle="fas" icon="file-download" /> Export
+            </bs.Button>
+            <bs.Button onClick={this.showImportModal.bind(this)}>
+              <FAIcon fastyle="fas" icon="file-upload" /> Import
+            </bs.Button>
+        </bs.ButtonToolbar>
+        <DeleteModal show={this.state.showDeleteModal} handleClose={this.handleDeleteModalClose.bind(this)} />
+        <ImportModal show={this.state.showImportModal} handleClose={this.handleImportModalClose.bind(this)} />
+      </>
     );
   }
 }
